Commit: Reorganizing files. Reorganizing PostController and PlaceController (moving mappings from post to PlaceController). Adding image saving option to EditPlace option
