+++ src/main/resources/application.properties
@@ -2,7 +2,7 @@ spring.datasource.url=jdbc:mysql://localhost:3306/hhu_db?serverTimezone=UTC
 #spring.datasource.driver-class-name=com.mysql.jdbc.Driver
 spring.datasource.username=root
 spring.datasource.password=password
-spring.jpa.hibernate.ddl-auto=create
+spring.jpa.hibernate.ddl-auto=validate
 
 # SQL dialect for generating optimized queries
 spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL5Dialect
