+++ src/main/resources/templates/editplace.html
@@ -1,0 +1,94 @@
+<!DOCTYPE html>
+<html lang="en" xmlns:th="http://www.thymeleaf.org" xmlns:sec="http://www.thymeleaf.org/extras/spring-security">
+<head>
+    <meta charset="UTF-8">
+    <title>HHU - Find Your Way</title>
+    <!-- Icon from: https://www.iconfinder.com/icons/345447/location_map_marker_navigation_icon -->
+    <link rel="icon" href="/images/logo.png" type="image/x-icon">
+    <link th:rel="stylesheet" th:href="@{webjars/bootstrap/4.6.0-1/css/bootstrap.css} "/>
+    <link th:rel="stylesheet" th:href="@{styles/map.css} "/>
+</head>
+<body>
+<nav class="navbar navbar-expand-lg navbar-dark bg-dark static-top">
+    <div class="container">
+        <a class="navbar-brand" href="/">HitchHikeUp</a>
+        <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarResponsive"
+                aria-controls="navbarResponsive"
+                aria-expanded="false" aria-label="Toggle navigation">
+            <span class="navbar-toggler-icon"></span>
+        </button>
+        <div class="collapse navbar-collapse" id="navbarResponsive">
+            <ul class="navbar-nav ml-auto">
+                <li class="nav-item active">
+                    <a class="nav-link" href="#">Home
+                        <span class="sr-only">(current)</span>
+                    </a>
+                </li>
+                <li class="nav-item">
+                    <a class="nav-link" href="#">About</a>
+                </li>
+                <li class="nav-item">
+                    <a class="nav-link" href="#">Contact</a>
+                </li>
+                <li class="nav-item">
+                    <a class="nav-link" href="/logout" sec:authorize="isAuthenticated()"> Logout</a>
+                </li>
+            </ul>
+        </div>
+    </div>
+</nav>
+<div id="before-map"
+     th:style="'background-image:url(' + @{/images/2.jpg} + '); height:100vh; background-position: center; background-repeat: no-repeat; background-size: cover;'">
+    <!-- Navigation -->
+    <div class="container">
+        <div id="editForm">
+            <form th:action="@{/place/edited}" th:method="post" th:object="${place}" th:autocomplete="off" enctype="multipart/form-data">
+                <input type="hidden" class="form-control" name="id" id="id" th:value="*{id}">
+                <div class="form-group row">
+                    <label for="name" class="col-sm-2 col-form-label" th:style="'background-color: white'">
+                        Name:
+                    </label>
+                    <div class="col-sm-6">
+                        <input type="text" class="form-control" name="name" id="name" th:value="*{name}">
+                    </div>
+                </div>
+                <!-- Auto wypelnienie wspolrzednymi z klikniecia na mapie -->
+                <input type="hidden" class="form-control" name="lat" id="latitude" th:value="*{lat}">
+                <input type="hidden" class="form-control" name="lng" id="longitude" th:value="*{lng}">
+                <div>
+                    <label>Direction </label>
+                    <select name="direction">
+                        <option th:value="*{direction}" th:text="*{direction}"></option>
+                        <option value="N">N</option>
+                        <option value="S">S</option>
+                        <option value="W">W</option>
+                        <option value="E">E</option>
+                        <option value="SW">SW</option>
+                        <option value="NW">NW</option>
+                        <option value="SE">SE</option>
+                        <option value="NE">NE</option>
+                    </select>
+                </div>
+                <div class="form-group row">
+                    <label for="description" class="col-sm-3 col-form-label" th:style="'background-color: white'">
+                        Description: </label>
+                    <div class="col-sm-10">
+                        <input type="text" class="form-control" name="description" id="description"
+                               th:value="*{description}">
+                    </div>
+                </div>
+                <div class="form-group row">
+                    <label for="photo" class="col-sm-3 col-form-label" th:style="'background-color: white'">
+                        Photo: </label>
+                    <div class="col-sm-10">
+                        <input type="file" class="form-control-file" name="photo" id="photo" >
+                        <img th:src="@{${'/place/show-photo-by-id/'+ place.id}}" class="float-right" style="max-width: 700px" alt="Photo of the place"/>
+                    </div>
+                </div>
+                <input type="submit" value="Add post" class="align-self-center btn-primary">
+            </form>
+        </div>
+    </div>
+</div>
+</body>
+</html>
