+++ src/main/resources/templates/index.html
@@ -44,14 +44,14 @@
         </div>
     </div>
 </nav>
-<div sec:authorize="!isAuthenticated()">
+<div sec:authorize="isAnonymous()">
     <div id="before-map"
          th:style="'background-image:url(' + @{/images/2.jpg} + '); height:55vh; background-position: center; background-repeat: no-repeat; background-size: cover;'">
         <!-- Navigation -->
         <div class="container">
             <div class="row">
                 <div class="col-sm-10 mt-5">
-                    <form method="post" sec:authorize="isAnonymous()" th:action="login">
+                    <form method="post"  th:action="login">
                         <div class="form-group row">
                             <label for="username" class="col-sm-2 col-form-label"
                                    th:style="'background-color: white'">Username:</label>
@@ -77,7 +77,7 @@
 </div>
 <div id="map"></div>
 <div id="postForm" class="hidden">
-    <form th:action="@{/post}" th:method="post" th:autocomplete="off" enctype="multipart/form-data">
+    <form th:action="@{/place/add}" th:method="post" th:autocomplete="off" enctype="multipart/form-data">
         <div class="form-group row">
             <label for="name" class="col-sm-2 col-form-label" th:style="'background-color: white'">
                 Name:
@@ -88,7 +88,7 @@
         </div>
         <!-- Auto wypelnienie wspolrzednymi z klikniecia na mapie -->
         <input type="hidden" class="form-control" name="lat" id="latitude" th:field="${place.lat}">
-        <input type="hidden" class="form-control" name="lng" id="longitude" th:field="*{place.lng}">
+        <input type="hidden" class="form-control" name="lng" id="longitude" th:field="${place.lng}">
         <div>
             <label>Direction </label>
             <select name="direction">
@@ -143,7 +143,7 @@
 
 <div id="addPostForm" class="hidden">
     <h3 id="placeName" class="text-center"></h3>
-    <form th:action="@{/addPost}" th:method="post" th:autocomplete="off">
+    <form th:action="@{/post/add}" th:method="post" th:autocomplete="off">
         <div class="form-group row">
             <label for="ratePost" class="col-form-label col-sm-3" th:style="'background-color: white'"> Rate
                 (max 10): </label>
@@ -171,10 +171,9 @@
 </div>
 
 <script defer
-        src="https://maps.googleapis.com/maps/api/js?key=YOUR_API_KEY&callback=initMap">
+        src="https://maps.googleapis.com/maps/api/js?key=APIKEY&callback=initMap">
 </script>
 <script th:inline="javascript"> var locationsList = [[${places}]]; </script>
-<!--<script th:inline="javascript"> var oneLocation= [[${place}]]; </script>-->
 
 <!-- deklaracja zmiennej globalnej - wartość pobierana z indexcontroller-->
 <script src="scripts/map.js"></script>
