+++ src/main/resources/templates/place.html
@@ -24,7 +24,7 @@
             <div class="collapse navbar-collapse" id="navbarResponsive">
                 <ul class="navbar-nav ml-auto">
                     <li class="nav-item active">
-                        <a class="nav-link" th:href="@{/hhu}">Home
+                        <a class="nav-link" th:href="@{/}">Home
                             <span class="sr-only">(current)</span>
                         </a>
                     </li>
